% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{ridgereg Class}
\description{
This class calculates all the interesting parameter for the ridge regression and uses the following methods:
}
\section{Fields}{

\describe{
\item{\code{beta}}{Regression Coefficients}

\item{\code{fitted_val}}{Fitted Values}

\item{\code{residual_val}}{Residual Values}

\item{\code{df}}{Degrees of Freedom}

\item{\code{res_var}}{Residual Variance}

\item{\code{var_reg}}{Variance of the Regression Coefficients}

\item{\code{t_values}}{t-values of the coefficients}

\item{\code{p_values}}{p-values of the coefficients}

\item{\code{formula}}{formula of the model}

\item{\code{call}}{Function Call as a string}
}}


