% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{visualize airport delays}
\usage{
visualize_airport_delays()
}
\value{
a \code{ggplot2} plot
}
\description{
This function visualizes the mean departure delay of different airports in US on the map
}
